Make build for api browser-friendly, re-org common and lexicon packages
