+++ packages/common/src/index.ts
@@ -6,7 +6,7 @@ export * from './fs'
 export * from './util'
 export * from './tid'
 export * from './ipld'
-export * from './ipld-multi'
+export * from './ipld-cbor'
 export * from './logger'
 export * from './types'
 export * from './streams'
