+++ packages/common/src/ipld-cbor.ts
@@ -1,0 +1,77 @@
+import { CID } from 'multiformats/cid'
+import * as Block from 'multiformats/block'
+import * as rawCodec from 'multiformats/codecs/raw'
+import { sha256 } from 'multiformats/hashes/sha2'
+import * as mf from 'multiformats'
+import * as cborx from 'cbor-x'
+import * as cborCodec from '@ipld/dag-cbor'
+import { check, schema } from '.'
+
+export const cborEncode = cborCodec.encode
+export const cborDecode = cborCodec.decode
+export const cborDecodeMulti = (encoded: Uint8Array): unknown[] => {
+  const decoded: unknown[] = []
+  cborx.decodeMultiple(encoded, (value) => {
+    decoded.push(value)
+  })
+  return decoded
+}
+
+export const dataToCborBlock = async (data: unknown) => {
+  return Block.encode({
+    value: data,
+    codec: cborCodec,
+    hasher: sha256,
+  })
+}
+
+export const cidForCbor = async (data: unknown): Promise<CID> => {
+  const block = await dataToCborBlock(data)
+  return block.cid
+}
+
+export const cborBytesToRecord = (
+  bytes: Uint8Array,
+): Record<string, unknown> => {
+  const val = cborDecode(bytes)
+  if (!check.is(val, schema.record)) {
+    throw new Error(`Expected object, got: ${val}`)
+  }
+  return val
+}
+
+export const verifyCidForBytes = async (cid: CID, bytes: Uint8Array) => {
+  const digest = await sha256.digest(bytes)
+  const expected = CID.createV1(cid.code, digest)
+  if (!cid.equals(expected)) {
+    throw new Error(
+      `Not a valid CID for bytes. Expected: ${expected.toString()} Got: ${cid.toString()}`,
+    )
+  }
+}
+
+export const sha256ToCid = (hash: Uint8Array, codec: number): CID => {
+  const digest = mf.digest.create(sha256.code, hash)
+  return CID.createV1(codec, digest)
+}
+
+export const sha256RawToCid = (hash: Uint8Array): CID => {
+  return sha256ToCid(hash, rawCodec.code)
+}
+
+// add extension for decoding CIDs
+// decoding code taken from @ipld/dag-cbor
+// does not support encoding cids
+cborx.addExtension({
+  Class: CID,
+  tag: 42,
+  encode: () => {
+    throw new Error('cannot encode cids')
+  },
+  decode: (bytes: Uint8Array): CID => {
+    if (bytes[0] !== 0) {
+      throw new Error('Invalid CID for CBOR tag 42; expected leading 0x00')
+    }
+    return CID.decode(bytes.subarray(1)) // ignore leading 0x00
+  },
+})
