+++ packages/common/src/ipld.ts
@@ -1,57 +1,7 @@
 import { CID } from 'multiformats/cid'
-import * as Block from 'multiformats/block'
-import * as rawCodec from 'multiformats/codecs/raw'
-import { sha256 } from 'multiformats/hashes/sha2'
-import * as mf from 'multiformats'
 import * as ui8 from 'uint8arrays'
-import * as cborCodec from '@ipld/dag-cbor'
 import { check, schema } from '.'
 import { z } from 'zod'
-
-export const cborEncode = cborCodec.encode
-export const cborDecode = cborCodec.decode
-
-export const dataToCborBlock = async (data: unknown) => {
-  return Block.encode({
-    value: data,
-    codec: cborCodec,
-    hasher: sha256,
-  })
-}
-
-export const cidForCbor = async (data: unknown): Promise<CID> => {
-  const block = await dataToCborBlock(data)
-  return block.cid
-}
-
-export const cborBytesToRecord = (
-  bytes: Uint8Array,
-): Record<string, unknown> => {
-  const val = cborDecode(bytes)
-  if (!check.is(val, schema.record)) {
-    throw new Error(`Expected object, got: ${val}`)
-  }
-  return val
-}
-
-export const verifyCidForBytes = async (cid: CID, bytes: Uint8Array) => {
-  const digest = await sha256.digest(bytes)
-  const expected = CID.createV1(cid.code, digest)
-  if (!cid.equals(expected)) {
-    throw new Error(
-      `Not a valid CID for bytes. Expected: ${expected.toString()} Got: ${cid.toString()}`,
-    )
-  }
-}
-
-export const sha256ToCid = (hash: Uint8Array, codec: number): CID => {
-  const digest = mf.digest.create(sha256.code, hash)
-  return CID.createV1(codec, digest)
-}
-
-export const sha256RawToCid = (hash: Uint8Array): CID => {
-  return sha256ToCid(hash, rawCodec.code)
-}
 
 export type JsonValue =
   | boolean
