+++ packages/pds/src/app-migrations/update-follow-subjects.ts
@@ -1,6 +1,7 @@
 import assert from 'assert'
 import { sql } from 'kysely'
 import { cborBytesToRecord, chunkArray } from '@atproto/common'
+import { cborToLexRecord } from '@atproto/repo'
 import AppContext from '../context'
 import Database from '../db'
 import { appMigration } from '../db/leader'
@@ -9,7 +10,6 @@ import { MessageDispatcher } from '../event-stream/message-queue'
 import { ids } from '../lexicon/lexicons'
 import { prepareUpdate, prepareDelete, assertValidRecord } from '../repo'
 import { RepoService } from '../services/repo'
-import { cborToLexRecord } from '@atproto/lexicon'
 
 const MIGRATION_NAME = '2023-03-14-update-follow-subjects'
 const SHORT_NAME = 'update-follow-subjects'
