+++ packages/pds/src/app-migrations/votes-to-likes.ts
@@ -1,6 +1,7 @@
 import assert from 'assert'
 import { sql } from 'kysely'
 import { cborBytesToRecord, chunkArray } from '@atproto/common'
+import { cborToLexRecord } from '@atproto/repo'
 import AppContext from '../context'
 import Database from '../db'
 import { appMigration } from '../db/leader'
@@ -9,7 +10,6 @@ import { ids } from '../lexicon/lexicons'
 import { prepareCreate, prepareDelete, assertValidRecord } from '../repo'
 import { RepoService } from '../services/repo'
 import { countAll } from '../db/util'
-import { cborToLexRecord } from '@atproto/lexicon'
 
 const MIGRATION_NAME = '2023-03-15-votes-to-likes'
 const SHORT_NAME = 'votes-to-likes'
