+++ packages/repo/src/block-map.ts
@@ -1,4 +1,5 @@
-import { lexToCborBlock, LexValue } from '@atproto/lexicon'
+import { dataToCborBlock } from '@atproto/common'
+import { lexToIpld, LexValue } from '@atproto/lexicon'
 import { CID } from 'multiformats/cid'
 import * as uint8arrays from 'uint8arrays'
 
@@ -6,7 +7,7 @@ export class BlockMap {
   private map: Map<string, Uint8Array> = new Map()
 
   async add(value: LexValue): Promise<CID> {
-    const block = await lexToCborBlock(value)
+    const block = await dataToCborBlock(lexToIpld(value))
     this.set(block.cid, block.bytes)
     return block.cid
   }
