+++ packages/repo/src/parse.ts
@@ -1,8 +1,9 @@
 import { check, cborDecode } from '@atproto/common'
-import { cborToLexRecord, RepoRecord } from '@atproto/lexicon'
+import { RepoRecord } from '@atproto/lexicon'
 import { CID } from 'multiformats/cid'
 import BlockMap from './block-map'
 import { MissingBlockError, UnexpectedObjectError } from './error'
+import { cborToLexRecord } from './util'
 
 export const getAndParseRecord = async (
   blocks: BlockMap,
