+++ packages/xrpc/package.json
@@ -18,7 +18,6 @@
   },
   "license": "MIT",
   "dependencies": {
-    "@atproto/common": "*",
     "@atproto/lexicon": "*",
     "zod": "^3.14.2"
   }
